requirement.txt
pandas==2.1.4
beautifulsoup4==4.12.2
lxml==5.0.0
watchdog==3.0.0
flask==3.0.0
flask-cors==4.0.0
werkzeug==3.0.1
pdfplumber==0.10.3
pymongo==4.6.1
mongomock==4.1.2
PyYAML==6.0.1
regex==2023.12.25
gpt4all==2.6.0
requests==2.31.0
